refactor: rename fs_utils to markdown_reader for clarity
